+++ test/it/fileutils/mountpoint_test.sh
@@ -1,0 +1,3 @@
+TestMountpointRoot() {
+    mountpoint /
+}