+++ test/it/spec/shred_spec.sh
@@ -1,0 +1,10 @@
+Describe 'shred'
+    Include fileutils/shred_test.sh
+    BeforeEach 'Setup'
+    AfterEach 'CleanUp'
+    It 'overwrites and removes the file'
+        When call TestShredRemove
+        The output should equal 'gone'
+        The status should be success
+    End
+End